Prevent dexlayout from moving code items if preceded by class data.
Class data has uleb encoded offsets to the code item, and moving code
items will change the size of the class data, which would in turn affect
the code item offsets if they precede the code items.

For now, make it so that dexlayout does not move the code items at all
in this case. A better fix would be to swap the order of the sections in
dexlayout, but that can be done in a future CL.

Unit test to follow.

Bug: 35855748
Test: mm test-art-host
Change-Id: Ica6da1e5e951cf0003fc9793f13ad10b74004eb9
(cherry picked from commit 863f1d79a70403443c089a91ca893db6cbadd3d8)