Fix 905 flake on low mem.
Test 905 could flake if a background GC occured. This could cause one
of the jvmtiEnvs to have its tag map updated and recomputed. Since
this map is unordered it leads to the order of free events being
different between the two environments. This is allowed (although
surprising) and test 905 assumed it was not the case.

We also expand this test to ensure that the right objects are getting
their free-events and that free events do not mess up the tagging of
other objects.

Test: ./test/testrunner/testrunner.py --host -j80
Bug: 119652731

Change-Id: Id17d1c87c710b80cd401dad109d98225ea28c29b
(cherry picked from commit 35a760d5e5879d50cd2d36e3d6286fe6c12888aa)